follow file standard and add human readable output on parse err
